Development

+++ CCDevTools/Controllers/ProjectsController.cs
@@ -7,29 +7,55 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using CCDevTools.Data;
 using CCDevTools.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
 
 namespace CCDevTools.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class ProjectsController : ControllerBase
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly IAuthorizationService _auth;
 
-        public ProjectsController(ApplicationDbContext context)
+        public ProjectsController(ApplicationDbContext context, UserManager<ApplicationUser> usr, IAuthorizationService auth)
         {
             _context = context;
+            _userManager = usr;
+            _auth = auth;
         }
 
         // GET: api/Projects
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Project>>> GetProjects()
         {
-          if (_context.Projects == null)
-          {
-              return NotFound();
-          }
-            return await _context.Projects.Include(p => p.Tickets).ToListAsync();
+            if (_context.Projects == null)
+            {
+                return NotFound();
+            }
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (userId == null)
+            {
+                return NotFound();
+            }
+
+            var membershipList = await _context.Memberships
+                .Where(m => m.UserId == userId)
+                .Select(m => m.ProjectId)
+                .ToListAsync();
+
+
+
+            return await _context.Projects
+                .Where(p => membershipList.Contains(p.Id))
+                .Include(p => p.Invitations)
+                .Include(p => p.Tickets).ToListAsync();
         }
 
         // GET: api/Projects/5
@@ -40,7 +66,11 @@ namespace CCDevTools.Controllers
           {
               return NotFound();
           }
-            var project = await _context.Projects.Include(p => p.Tickets).Where(p => p.Id == id).FirstOrDefaultAsync();
+            var project = await _context.Projects
+                .Include(p => p.Tickets)
+                .Include(p => p.Invitations)
+                .Where(p => p.Id == id)
+                .FirstOrDefaultAsync();
 
             if (project == null)
             {
@@ -58,6 +88,14 @@ namespace CCDevTools.Controllers
             if (id != project.Id)
             {
                 return BadRequest();
+            }
+
+            // check if we are able to modify the entry (maintainer or better, level 10 or lower)
+            AuthorizationResult result = await _auth.AuthorizeAsync(User, project, "IsMaintainer");
+
+            if (!result.Succeeded)
+            {
+                return Unauthorized();
             }
 
             _context.Entry(project).State = EntityState.Modified;
@@ -90,10 +128,31 @@ namespace CCDevTools.Controllers
           {
               return Problem("Entity set 'ApplicationDbContext.Projects'  is null.");
           }
+
+          // TODO: Add a membership to the new project for the user as an owner
+          var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+            
             _context.Projects.Add(project);
             await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetProject", new { id = project.Id }, project);
+            if (project.Id > 0)
+            {
+                _context.Memberships.Add(new Membership
+                {
+                    UserId = userId,
+                    ProjectId = project.Id,
+                    Level = 0
+                });
+            }
+
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction("GetProject", new { id = project.Id });
         }
 
         // DELETE: api/Projects/5
@@ -104,10 +163,19 @@ namespace CCDevTools.Controllers
             {
                 return NotFound();
             }
+
             var project = await _context.Projects.FindAsync(id);
             if (project == null)
             {
                 return NotFound();
+            }
+
+            // only allow if the user is an owner of the project
+            AuthorizationResult result = await _auth.AuthorizeAsync(User, project, "IsOwner");
+
+            if (!result.Succeeded)
+            {
+                return Unauthorized();
             }
 
             _context.Projects.Remove(project);
